core: Move SandboxType from avm1 to sandbox module
